+++ packages/html/src/define/media-current-time-display.ts
@@ -1,0 +1,4 @@
+import { CurrentTimeDisplayElement } from '@/elements/current-time-display';
+import { defineCustomElement } from '@/utils/custom-element';
+
+defineCustomElement('media-current-time-display', CurrentTimeDisplayElement);