+++ packages/html/src/define/media-duration-display.ts
@@ -1,0 +1,4 @@
+import { DurationDisplayElement } from '@/elements/duration-display';
+import { defineCustomElement } from '@/utils/custom-element';
+
+defineCustomElement('media-duration-display', DurationDisplayElement);