+++ packages/html/src/define/media-popover.ts
@@ -1,0 +1,4 @@
+import { PopoverElement } from '@/elements/popover';
+import { defineCustomElement } from '@/utils/custom-element';
+
+defineCustomElement('media-popover', PopoverElement);