+++ packages/html/src/icons/media-pause-icon.ts
@@ -1,5 +1,6 @@
 import { SVG_ICONS } from '@videojs/icons';
 
+import { defineCustomElement } from '@/utils/custom-element';
 import { MediaChromeIcon } from './media-chrome-icon';
 
 export function getTemplateHTML() {
@@ -9,8 +10,8 @@ export function getTemplateHTML() {
   `;
 }
 
-export class MediaPauseIcon extends MediaChromeIcon {
+export class MediaPauseIconElement extends MediaChromeIcon {
   static getTemplateHTML: () => string = getTemplateHTML;
 }
 
-customElements.define('media-pause-icon', MediaPauseIcon);
+defineCustomElement('media-pause-icon', MediaPauseIconElement);