+++ packages/html/src/media/media-provider.ts
@@ -4,15 +4,12 @@ import type { MediaStore } from '@videojs/core/store';
 import { ProviderMixin } from '@open-wc/context-protocol';
 import { createMediaStore } from '@videojs/core/store';
 
-const ProviderHTMLElement: Constructor<CustomElement> = ProviderMixin(HTMLElement);
+const ProviderHTMLElement: Constructor<CustomElement & HTMLElement> = ProviderMixin(HTMLElement);
 
-export class MediaProvider extends ProviderHTMLElement {
+export class MediaProviderElement extends ProviderHTMLElement {
   contexts = {
     mediaStore: (): MediaStore => {
       return createMediaStore();
     },
   };
 }
-
-// @ts-expect-error - fix types after (Rahim)
-customElements.define('media-provider', MediaProvider);