+++ site/src/examples/html/mute-button/mute-button.js
@@ -1,4 +1,4 @@
-import '@videojs/html';
+import '@videojs/html/define/media-mute-button';
 import '@videojs/html/icons';
 
 // The web components will automatically register themselves