+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.0-preview.4](https://github.com/videojs/v10/compare/@videojs/utils@0.1.0-preview.3...@videojs/utils@0.1.0-preview.4) (2025-10-30)
+
+
+### Miscellaneous Chores
+
+* **@videojs/utils:** Synchronize videojs versions
+
 ## [0.1.0-preview.3](https://github.com/videojs/v10/compare/@videojs/utils@0.1.0-preview.2...@videojs/utils@0.1.0-preview.3) (2025-10-29)